backend/pyproject.toml
[project]
name = "myglot"
version = "0.1.0"
description = "Personal language learning app"
requires-python = ">=3.11"
dependencies = [
    "google-cloud-translate>=3.15",
    "google-cloud-texttospeech>=2.16",
    "google-cloud-speech>=2.26",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "sqlmodel>=0.0.19",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "httpx>=0.27"]
lint = ["ruff>=0.11", "mypy>=1.14"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["app"]

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "SIM", "RUF"]
ignore = ["E501", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.lint.per-file-ignores]
"app/models.py" = ["UP006", "UP035", "UP045"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
